+++ Arc/src/Arc/Renderer/Framebuffer.h
@@ -66,8 +66,8 @@ namespace ArcEngine
 
 		virtual void Resize(uint32_t width, uint32_t height) = 0;
 		
-		virtual uint32_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
-		virtual uint32_t GetDepthAttachmentRendererID() const = 0;
+		virtual uint64_t GetColorAttachmentRendererID(uint32_t index = 0) const = 0;
+		virtual uint64_t GetDepthAttachmentRendererID() const = 0;
 		
 		virtual const FramebufferSpecification& GetSpecification() const = 0;
