+++ Arc/src/Platform/OpenGL/OpenGLFramebuffer.h
@@ -20,10 +20,11 @@ namespace ArcEngine
 
 		virtual void Resize(uint32_t width, uint32_t height) override;
 		
-		virtual uint32_t GetColorAttachmentRendererID(uint32_t index = 0) const override { ARC_CORE_ASSERT(index < m_ColorAttachments.size()); return m_ColorAttachments[index]; }
-		virtual uint32_t GetDepthAttachmentRendererID() const override { return m_DepthAttachment; }
+		virtual uint64_t GetColorAttachmentRendererID(uint32_t index = 0) const override { ARC_CORE_ASSERT(index < m_ColorAttachments.size()); return m_ColorAttachments[index]; }
+		virtual uint64_t GetDepthAttachmentRendererID() const override { return m_DepthAttachment; }
 		
 		virtual const FramebufferSpecification& GetSpecification() const override { return m_Specification; }
+
 	private:
 		uint32_t m_RendererID = 0;
 		FramebufferSpecification m_Specification;
